feat: update tasks
