+++ package.json
@@ -16,10 +16,10 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "test:1_task": "react-scripts test src/tests/1_task.test.js",
-    "test:2_task": "react-scripts test src/tests/2_task.test.js",
-    "test:3_task": "react-scripts test src/tests/3_task.test.js",
-    "test:4_task": "react-scripts test src/tests/4_task.test.js",
+    "test:1": "react-scripts test src/tests/1_task.test.js",
+    "test:2": "react-scripts test src/tests/2_task.test.js",
+    "test:3": "react-scripts test src/tests/3_task.test.js",
+    "test:4": "react-scripts test src/tests/4_task.test.js",
     "test:all": "react-scripts test src/tests"
   },
   "eslintConfig": {
