+++ src/solutions/3_task.js
@@ -1,3 +1,12 @@
 // Please implement your solution in this file
 
+export const LOADING_MESSAGE = "Rockets loading...";
+
+const FILTER_PARAMS = {
+  year: 2018,
+  customerName: "NASA",
+};
+
 export const RocketItem = ({ rocket }) => <div>RocketItem</div>;
+
+export const RocketsList = () => <div>RocketsList</div>;
