--- src/solutions/4_task.js
@@ -1,8 +1,0 @@
-// Please implement your solution in this file
-
-const FILTER_PARAMS = {
-  year: 2018,
-  customerName: "NASA",
-};
-
-export const RocketsList = () => <div>RocketsList</div>;
